// frontend/src/app/api/models/event-page.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { CrashWatch } from '../models/crash-watch';
import { DashboardEvent } from '../models/dashboard-event';
import { QuietCount } from '../models/quiet-count';
export interface EventPage {
  crashWatch: CrashWatch;
  events: Array<DashboardEvent>;
  nextBefore: (number | null);
  quiet: Array<QuietCount>;
  quietSince: string;
}